feat: use new sqlite cache for uni v2 events (#383)
* feat: use Events to init uniswap

* fix: circular import

* fix: circular import

* fix: circular import

* fix: args should be kwargs

* fix: circular import

* fix: TypeError

* fix: RecursionError

* feat: use Events class for uni v3

* fix: AttributeError

* fix: EventLookupError

* fix: EventLookupError

* fix: Contract.events

* fix: Contract.events

* fix: Contract.events

* fix: does this fit it?

* fix: contract Events topics

* fix: semaphore locked

* feat: batch size env

* feat: default chunks per batch disabled

* fix: cache metadata update

* feat: use async provider for getLogs

* feat: insert chunks in background

* fix: db.set_price for eth

* fix: if check

* fix: cache metadata

* fix: database is locked

* fix: database is locked

* fix: database is locked

* fix: database is locked

* fix: database is locked

* feat: retry_locked deco

* fix: database is locked

* feat: apply @retry_locked to get_token, get_price, set_price

* fix: use eth_portfolio db utils if present

* fix: _load_cache

* fix: use eth-port db utils where possible

* fix: oops

* fix: ImportError

* fix: force await db_task

* fix: NoneType not awaitable

* fix: force insert each chunk

* fix: cache_metadata

* fix: db.set_price

* feat: debug loggers

* chore: refactor

* fix: forgot to delete

* fix: name conflict

* fix: get get token

* fix: log topic indexing in db

* feat: more indexes for logs

* feat: log select sql

* fix: skip DISTINCT in log lookup

* fix: retry InterfaceError

* fix: better query specs

* fix: better query specs

* feat: paginate results

* fix: order db results by block

* fix: order db results by block

* chore: refactor

* fix: async -> sync

* chore: refactor

* fix: syntax err

* fix: yearn TypeError

* fix: name conflict

* feat: add exception details

* fix: short results

* fix: Contract.events

* fix: wsteth type error

* fix: UnexpectedError db is locked

* fix: mixing txs

* fix: price precision

* chore: remove unneeded assert

* fix: missing address check

* fix: use _get_get_token for token params

* chore: refactor

* fix: wsteth type error

* fix: curve TypeError

* fix: db_session required

* fix: pools topics

* fix: query metadata

* fix: WeiBalance.value_usd

* fix: TypeError

* fix: incorrect number of results

* fix: remove pagination

* fix: logs cache check for addresses if single addr

* fix: set_metadata

* fix: accidental deletion

* fix: set_metadata for single address

* feat: ossify db

* fix: step down batch size if needed

* feat: clean_addresses

* fix: TypeError

* feat: YPRICEMAGIC_SKIP_CACHE env

* fix: GETLOGS_BATCH_SIZE env

* feat: use ProcessedEvents for uni v2

* fix: import err

* feat: yield now, insert later

* feat: LogFilter __repr__

* feat: debug loggers

* feat: better logger

* feat: better logger

* fix: exit check

* fix: end < start edge case

* fix: AttributeError

* feat: insert debug loggers

* fix: yearn type err

* feat: get_pools_for ttl

* fix: get_pools_for

* fix: async

* feat: helper methods

* fix: is_cached_thru for fitlers with address

* fix: _load_new_objects

* fix: db envs

* chore: remove old imports

* fix: db envs

+++ y/prices/stable_swap/curve.py
@@ -1,6 +1,7 @@
 import asyncio
 import logging
 from collections import defaultdict
+from decimal import Decimal
 from enum import IntEnum
 from functools import cached_property
 from typing import Dict, List, NoReturn, Optional, Tuple
@@ -137,7 +138,7 @@ class CurvePool(ERC20): # this shouldn't be ERC20 but works for inheritance for 
         return [ERC20(coin, asynchronous=self.asynchronous) for coin in coins if coin != ZERO_ADDRESS]
     
     @a_sync.a_sync(ram_cache_maxsize=1000)
-    async def get_balances(self, block: Optional[Block] = None) -> Dict[ERC20, int]:
+    async def get_balances(self, block: Optional[Block] = None) -> Dict[ERC20, Decimal]:
         """
         Get {token: balance} of liquidity in the pool.
         """
@@ -164,7 +165,7 @@ class CurvePool(ERC20): # this shouldn't be ERC20 but works for inheritance for 
             raise ValueError(f'could not fetch balances {self.__str__()} at {block}')
 
         return {
-            coin: balance / 10 ** dec
+            coin: Decimal(balance / 10 ** dec)
             for coin, balance, dec in zip(coins, balances, decimals)
             if coin != ZERO_ADDRESS
         }
@@ -191,7 +192,7 @@ class CurvePool(ERC20): # this shouldn't be ERC20 but works for inheritance for 
         prices = await asyncio.gather(*[coin.price(block=block, sync=False) for coin in balances])
 
         return UsdValue(
-            sum(balance * price for balance, price in zip(balances.values(),prices))
+            sum(balance * Decimal(price) for balance, price in zip(balances.values(), prices))
         )
     
     @a_sync.a_sync(ram_cache_maxsize=10_000, ram_cache_ttl=10*60)
